requirements.txt
ipwhois
pyopenssl
shodan
lxml
python-dotenv
requests[socks]
